Move RELATIVE_IMPORT_RE to packages/cli/src/lib/patterns.ts
A dependency-free module so any site build script can import the regex
without pulling in heavy transitive deps like @barefootjs/jsx.
