+++ site/ui/__tests__/inline-relative-imports.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from 'bun:test'
-import { RELATIVE_IMPORT_RE } from '../build-utils'
+import { RELATIVE_IMPORT_RE } from '../../../packages/cli/src/lib/patterns'
 
 function matchAll(input: string) {
   return [...input.matchAll(RELATIVE_IMPORT_RE)]
